feature: add ISQ7000, Orbitrap Exploris GC 240, support multiple FAIMS voltages (#26)
* Add Orbitrap Exploris GC 240, to supported instruments. Added some debugging statements to make diagnosing missing instrument types easier in the future

* Add ISQ 7000, fix typo in reverse_transpose

* remove unknown detector type, bump number of instruments

* support multiple CompensationVoltageValues

* Add handling of FAIMS start and end ramp, add warning if there are more than 2 faims voltages

* Update Cargo.toml

---------

+++ src/spectrum/bindata/encodings.rs
@@ -140,11 +140,11 @@ mod byte_rotation {
     }
 
     pub fn reverse_transpose_i64(data: &[u8]) -> Vec<u8> {
-        reverse_transpose_8bytes::<f64>(data)
+        reverse_transpose_8bytes::<i64>(data)
     }
 
     pub fn reverse_transpose_f64(data: &[u8]) -> Vec<u8> {
-        reverse_transpose_8bytes::<i64>(data)
+        reverse_transpose_8bytes::<f64>(data)
     }
 }
